Add abstracted class
